Try1/requirements.txt
Django==2.1.5
django-jet==1.0.8
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
wincertstore==0.2
